MAILBOX-211 Moving event related interfaces and code to a new package
