+++ mailbox/store/src/main/java/org/apache/james/mailbox/store/search/ListeningMessageSearchIndex.java
@@ -28,10 +28,10 @@ import org.apache.james.mailbox.MailboxSession;
 import org.apache.james.mailbox.exception.MailboxException;
 import org.apache.james.mailbox.model.MessageRange;
 import org.apache.james.mailbox.model.UpdatedFlags;
-import org.apache.james.mailbox.store.MailboxEventDispatcher.AddedImpl;
-import org.apache.james.mailbox.store.MailboxEventDispatcher.ExpungedImpl;
-import org.apache.james.mailbox.store.MailboxEventDispatcher.FlagsUpdatedImpl;
-import org.apache.james.mailbox.store.MailboxEventDispatcher.MailboxDeletionImpl;
+import org.apache.james.mailbox.store.event.MailboxEventDispatcher.AddedImpl;
+import org.apache.james.mailbox.store.event.MailboxEventDispatcher.ExpungedImpl;
+import org.apache.james.mailbox.store.event.MailboxEventDispatcher.FlagsUpdatedImpl;
+import org.apache.james.mailbox.store.event.MailboxEventDispatcher.MailboxDeletionImpl;
 import org.apache.james.mailbox.store.mail.MessageMapper.FetchType;
 import org.apache.james.mailbox.store.mail.MessageMapperFactory;
 import org.apache.james.mailbox.store.mail.model.MailboxId;
